+++ espresso/src/com.oracle.truffle.espresso.staticobject/src/com/oracle/truffle/espresso/staticobject/DefaultStaticProperty.java
@@ -25,8 +25,8 @@ package com.oracle.truffle.espresso.staticobject;
 public final class DefaultStaticProperty extends StaticProperty {
     private final String id;
 
-    public DefaultStaticProperty(String id, StaticPropertyKind kind, boolean isFinal) {
-        super(kind, isFinal);
+    public DefaultStaticProperty(String id, StaticPropertyKind kind, boolean storeAsFinal) {
+        super(kind, storeAsFinal);
         this.id = id;
     }