+++ substratevm/src/com.oracle.svm.reflect/src/com/oracle/svm/reflect/hosted/ReflectionFeature.java
@@ -38,6 +38,7 @@ import com.oracle.svm.core.configure.ReflectionConfigurationParser;
 import com.oracle.svm.core.graal.GraalFeature;
 import com.oracle.svm.hosted.FallbackFeature;
 import com.oracle.svm.hosted.FeatureImpl.DuringSetupAccessImpl;
+import com.oracle.svm.hosted.FeatureImpl.FeatureAccessImpl;
 import com.oracle.svm.hosted.ImageClassLoader;
 import com.oracle.svm.hosted.analysis.Inflation;
 import com.oracle.svm.hosted.config.ConfigurationParserUtils;
@@ -60,6 +61,9 @@ public final class ReflectionFeature implements GraalFeature {
     public void afterRegistration(AfterRegistrationAccess access) {
         ModuleSupport.exportAndOpenPackageToUnnamed("java.base", "jdk.internal.reflect", false);
         ModuleSupport.openModuleByClass(ReflectionProxy.class, null);
+
+        reflectionData = new ReflectionDataBuilder((FeatureAccessImpl) access);
+        ImageSingletons.add(RuntimeReflectionSupport.class, reflectionData);
     }
 
     @Override
@@ -70,9 +74,6 @@ public final class ReflectionFeature implements GraalFeature {
         ReflectionSubstitution subst = new ReflectionSubstitution(access.getMetaAccess().getWrapped(), access.getHostVM().getClassInitializationSupport(), access.getImageClassLoader());
         access.registerSubstitutionProcessor(subst);
         ImageSingletons.add(ReflectionSubstitution.class, subst);
-
-        reflectionData = new ReflectionDataBuilder(access);
-        ImageSingletons.add(RuntimeReflectionSupport.class, reflectionData);
 
         access.registerObjectReplacer(new ReflectionObjectReplacer(access.getMetaAccess()));