+++ sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/PlatformCapability.java
@@ -68,6 +68,11 @@ public abstract class PlatformCapability<S extends Enum<S> & LLVMSyscallEntry> i
         valueToSysCall = initTable(cls);
     }
 
+    public void initializeThread(@SuppressWarnings("unused") LLVMContext context,
+                    @SuppressWarnings("unused") Thread thread) {
+        // Nothing needs to be done in Sulong for native thread initialization.
+    }
+
     @SuppressWarnings("unchecked")
     private S[] initTable(Class<S> cls) {
         S[] constants = cls.getEnumConstants();